+++ topl-btc-bridge/src/main/scala/co/topl/bridge/package.scala
@@ -5,6 +5,33 @@ import _root_.quivr.models.Int128
 package object bridge {
 
   class BTCWaitExpirationTime(val underlying: Int) extends AnyVal
+  class ToplWaitExpirationTime(val underlying: Int) extends AnyVal
+
+  object implicits {
+    // implicit class for BTCWaitExpirationTime that implements <, >, <=, >=, ==, != with integers
+    implicit class BTCWaitExpirationTimeOps(
+        val btcWaitExpirationTime: BTCWaitExpirationTime
+    ) {
+      def <(other: Long): Boolean = btcWaitExpirationTime.underlying < other
+      def >(other: Long): Boolean = btcWaitExpirationTime.underlying > other
+      def <=(other: Long): Boolean = btcWaitExpirationTime.underlying <= other
+      def >=(other: Long): Boolean = btcWaitExpirationTime.underlying >= other
+      def ==(other: Long): Boolean = btcWaitExpirationTime.underlying == other
+      def !=(other: Long): Boolean = btcWaitExpirationTime.underlying != other
+    }
+
+    // implicit class for ToplWaitExpirationTime that implements <, >, <=, >=, ==, != with integers
+    implicit class ToplWaitExpirationTimeOps(
+        val toplWaitExpirationTime: ToplWaitExpirationTime
+    ) {
+      def <(other: Long): Boolean = toplWaitExpirationTime.underlying < other
+      def >(other: Long): Boolean = toplWaitExpirationTime.underlying > other
+      def <=(other: Long): Boolean = toplWaitExpirationTime.underlying <= other
+      def >=(other: Long): Boolean = toplWaitExpirationTime.underlying >= other
+      def ==(other: Long): Boolean = toplWaitExpirationTime.underlying == other
+      def !=(other: Long): Boolean = toplWaitExpirationTime.underlying != other
+    }
+  }
 
   class Fellowship(val underlying: String) extends AnyVal